match exact variable name in update
This applies a stricter search criteria when updating lines in the nml. Before, the line would be updated if the line contained the variable to be updated, e.g. if the variable to be updated is "days", the "days_to_increment" would also be updated.

The stricter search criteria ensures that only the whole variable matches.